+++ next_action/__init__.py
@@ -1,3 +1,4 @@
 from .pick_action import next_action_based_on_priority
 
+__name__ = "next-action"
 __version__ = "0.0.3"